\name{ABPvirus}
\docType{data}
\alias{ABPvirus}
\title{Data on Acute Bee Paralysis Virus Infections in Bumblebees from Pascall et al. (2018).}
\description{Acute Bee Paralysis Virus infection status in Bumblebees from pooled samples (Pascall et al. (2018)).}
\usage{data(ABPvirus)}
\format{A data frame with 99 rows and 3 columns. Pooled samples of bumblebees, conisting of \code{number} individuals, were assayed for an Acute Bee Paralysis virus infection. In samples where the virus was detected \code{infected} is equal to one, or zero otherwise. The column \code{species} gives the species' latin name (the genus is \emph{Bombus}).}
\references{Pascall, D.J. et al. 2018 BioRxiv 498717}  
\keyword{datasets}
