\name{BTdata}
\docType{data}
\alias{BTdata}
\title{Blue Tit Data for a Quantitative Genetic Experiment}
\description{Blue Tit (\emph{Cyanistes caeruleus}) Data for a Quantitative Genetic Experiment}
\usage{data(BTdata)}
\format{a data frame with 828 rows and 7 columns, with variables tarsus length (\code{tarsus}) and colour (\code{back}) measured on 828 individuals (\code{animal}). The mother of each is also recorded (\code{dam}) together with the foster nest (\code{fosternest}) in which the chicks were reared. The date on which the first egg in each nest hatched (\code{hatchdate}) is recorded together with the sex (\code{sex}) of the individuals.}
\references{
Hadfield, J.D. et. al. 2007 Journal of Evolutionary Biology 20 549-557
}
\seealso{\link{BTped}}
\keyword{datasets}
