\name{BTped}
\docType{data}
\alias{BTped}
\title{Blue Tit Pedigree}
\description{Blue Tit (\emph{Cyanistes caeruleus}) Pedigree}
\usage{data(BTped)}
\format{a data frame with 1040 rows and 3 columns, with individual identifier (\code{animal}) mother identifier (\code{dam}) and father identifier (\code{sire}). The first 212 rows are the parents of the 828 offspring from 106 full-sibling families. Parents are assumed to be unrelated to each other and have NA's in the \code{dam} and \code{sire} column. 
}
\references{
Hadfield, J.D. et. al. 2007 Journal of Evolutionary Biology 20 549-557
}
\seealso{\link{BTped}}
\keyword{datasets}


