\name{FullGrumpy}
\docType{data}
\alias{FullGrumpy}
\title{Full Grumpy Scores of Academics}
\description{Ordinal ratings of grumpiness as assessed from photos of academics.}
\usage{data(FullGrumpy)}
\format{A data frame with 5,368 rows and 8 columns. Two photos (\code{photo}) of 22 people (\code{person}) working in the Institute of Evolution and Ecology, Edinburgh, were taken. In one photo the person was happy and in the other they were grumpy (\code{type}).  122 \code{respondents} gave a score between 1 and 10 (\code{y}) indicating how grumpy they thought each person looked in each photo (with 10 being the most grumpy). The \code{age} of the person photographed and the number of years between when they published their first academic paper and the photo was taken (\code{ypub}) were recorded. Whether the \code{respondent} was a \code{student} in the institute was also recorded.}  
\keyword{datasets}
