\name{Grumpy}
\docType{data}
\alias{Grumpy}
\title{Grumpy Scores of Academics}
\description{Ordinal ratings of grumpiness as assessed from photos of academics.}
\usage{data(Grumpy)}
\format{A data frame with 44 rows and 8 columns. Two photos (\code{photo}) of 22 people (\code{person}) working in the Institute of Evolution and Ecology, Edinburgh, were taken. In one photo the person was happy and in the other they were grumpy (\code{type}).  122 respondents gave a score between 1 and 10 indicating how grumpy they thought each person looked in each photo (with 10 being the most grumpy). \code{y} gives the average score given by the 122 respondents. The number of respondents giving a photo a score of 5 or less (\code{l5}) or more than 5 (\code{g5}) was also recorded in addition to the person's age (\code{age}) and the number of years between when they published their first academic paper and the photo was taken (\code{ypub}).}  
\keyword{datasets}
