\name{KPPM}
\alias{KPPM}
\title{Kronecker Product Permutation Matrix}
\description{
  Forms an mk x mk Kronecker Product Permutation Matrix
}
\usage{
KPPM(m, k)
}
\arguments{
  \item{m}{integer}
  \item{k}{integer}
}
\value{
  \item{}{Matrix}
}
\references{Schott, J.R.(2003) Journal of Multivariate Analysis 87 (1) 177-190}
\author{Jarrod Hadfield \email{j.hadfield@ed.ac.uk}}
\examples{
KPPM(2,3)
}
\keyword{array}
