\name{PlodiaPO}
\docType{data}
\alias{PlodiaPO}
\title{Phenoloxidase measures on caterpillars of the Indian meal moth.}
\description{Phenoloxidase measures on caterpillars of the Indian meal moth 
(\emph{Plodia interpunctella}).}
\usage{PlodiaPO}
\format{a data frame with 511 rows and 3 columns, with variables indicating 
full-sib family (\code{FSfamily}), phenoloxidase measures (\code{PO}), and plate 
(\code{plate}). PO has undergone a Box-Cox power transformation of 0.141}
\source{Tidbury H & Boots M (2007) University of Sheffield}
\seealso{\link{PlodiaR}, \link{PlodiaRB}}
\keyword{datasets}


