\name{PlodiaR}
\docType{data}
\alias{PlodiaR}
\title{Resistance of Indian meal moth caterpillars to the granulosis virus 
PiGV.}
\description{Resistance of Indian meal moth (\emph{Plodia interpunctella}) 
caterpillars to the granulosis virus PiGV.}
\usage{PlodiaR}
\format{a data frame with 50 rows and 5 columns, with variables indicating full-
sib family (\code{FSfamly}), date of egg laying (\code{date_laid}) and assaying 
(\code{date_Ass}), and the number of individuals from the family that were 
experimentally infected with the virus \code{Infected} and the number of those 
that pupated \code{Pupated}. These full-sib family identifiers also relate to 
the full-sib family identifiers in PlodiaPO}
\source{Tidbury H & Boots M (2007) University of Sheffield}
\seealso{\link{PlodiaRB}, \link{PlodiaPO}}
\keyword{datasets}


