\name{PlodiaRB}
\docType{data}
\alias{PlodiaRB}
\title{Resistance (as a binary trait) of Indian meal moth caterpillars to the 
granulosis virus PiGV.}
\description{Resistance (as a binary trait) of Indian meal moth (\emph{Plodia 
interpunctella}) caterpillars to the granulosis virus PiGV.}
\usage{data(PlodiaRB)}
\format{a data frame with 784 rows and 4 columns, with variables indicating full-
sib family (\code{FSfamly}), date of egg laying (\code{date_laid}) and assaying 
(\code{date_Ass}), and a binary variable indicating whether an individual was 
resistant (\code{Pupated}) to an experimental infection of the virus. These data 
are identical to those in the data.frame \code{PlodiaR} except each family-level 
binomial variable has been expanded into a binary variable for each individual.}
\source{Tidbury H & Boots M (2007) University of Sheffield}
\seealso{\link{PlodiaR}, \link{PlodiaPO}}
\keyword{datasets}


