\name{plot.MCMCglmm}
\alias{plot.MCMCglmm}
\title{Plots MCMC chains from MCMCglmm using plot.mcmc}
\description{\code{plot} method for class \code{"MCMCglmm"}.
}
\usage{
\method{plot}{MCMCglmm}(x, random=FALSE, \dots)
}
\arguments{
 \item{x}{an object of class \code{"MCMCglmm"}}
 \item{random}{logical; should saved random effects be plotted}
 \item{\dots}{Further arguments to be passed}
}
\author{Jarrod Hadfield \email{j.hadfield@ed.ac.uk}}
\seealso{\code{plot.mcmc}, \code{\link{MCMCglmm}}}
\keyword{hplot}

