\name{sigma}
\docType{data}
\alias{sigma}
\title{Infection Data for Sigma Virus in \emph{Drosophila melanogaster}}
\description{Infection data for sigma virus isolates across lines of \emph{Drosophila melanogaster}}
\usage{data(sigma)}
\format{a data frame with 1092 rows and 4 columns. \code{infected} and \code{not_infected} give the number of flies in a vial that are infected or not infected with sigma virus. The viruses belong to one of five strains designated by their country of origin: United States (AP30), Spain (GC20), Greece (PF115), the United Kingdom (E27) and France (Hap23). The flies are genetically identical except at the third chromosome for which they belong to one of 67 lines \code{line}.}
\references{
Carpenter, J.A et al. 2012 Evolution 66 1114--1125
}
\seealso{\link{sigma}}
\keyword{datasets}
