\name{SShorns}
\docType{data}
\alias{SShorns}
\title{Horn type and genders of Soay Sheep}
\description{Horn type and genders of Soay Sheep \emph{Ovis aires}}
\usage{BTdata}
\format{a data frame with 666 rows and 3 columns, with individual idenitifier (\code{id}), horn type (\code{horn}) and gender (\code{sex}).}
\references{
Clutton-Brock T., Pemberton, J. Eds. 2004 Soay Sheep: Dynamics \& Selection in an Island Population
}
\keyword{datasets}
