\name{tufted_puffin}
\docType{data}
\alias{tufted_puffin}
\title{Diet composition of Tufted Puffin}
\description{Diet composition of Tufted Puffin (\emph{Fratercula cirrhata}) chicks from Middleton Island in the Gulf of Alaska (1978-2024).}
\usage{data(tufted_puffin)}
\format{a data frame with 565 rows and 7 columns. Each row corresponds to a sample and \code{year} and \code{month} refer to the date of collection. The following four columns are the number of Capelin (\code{Capelin}), Pacific Herring (\code{Herring}), Walleye pollock (\code{Pollock}) and Pacific Sand lance (\code{SandLance}) in each sample. An additinoal 40 prey taxa were also recorded and their combined counts per sample are in the final column (\code{Other}).}
\source{Hatch, S., Arimitsu, M., Piatt, J., Whelan, S., and Marsteller, C.E., 2023, Seabird Diet Data Collected on Middleton Island, Gulf of Alaska (ver 3.0, February 2025): U.S. Geological Survey data release, \url{https://doi.org/10.5066/P93I0P67}.}
\keyword{datasets}